Trait tvserver::domain::traits::MediaStorer

pub trait MediaStorer: Send + Sync {
    fn list<'life0, 'life1, 'async_trait>(
        &'life0 self,
        collection: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<VideoEntry, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn move_file<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn rename<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        current: &'life1 str,
        new_name: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn as_path(&self, collection: &str, video: &str) -> String;
    fn convert_to_mp4<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

Interface to a repository of available media files, currently implemented for the file system but could also support an S3 object store for instance.

Required Methods

fn list<'life0, 'life1, 'async_trait>(
&'life0 self,
collection: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<VideoEntry, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn move_file<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn rename<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
current: &'life1 str,
new_name: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

fn delete<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn as_path(&self, collection: &str, video: &str) -> String

fn convert_to_mp4<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,