Trait tvserver::domain::traits::TaskMonitor

pub trait TaskMonitor: Sync + Send {
    fn get_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TaskState> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_key(&self) -> String;
    fn get_seconds_since_finished(&self) -> i64;
    fn terminate(&self);
    fn has_finished(&self) -> bool;
    fn cleanup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        store: &'life1 Arc<dyn MediaStorer>,
        force_delete: bool
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

Provides a common interface to obtain the state of a task and terminate it.

In this context, a task is a long running task, like a downloading or converting a video

Required Methods

fn get_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = TaskState> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

fn get_key(&self) -> String

fn get_seconds_since_finished(&self) -> i64

fn terminate(&self)

fn has_finished(&self) -> bool

fn cleanup<'life0, 'life1, 'async_trait>(
&'life0 self,
store: &'life1 Arc<dyn MediaStorer>,
force_delete: bool
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,